requirements.txt
playwright
pytest
pyyaml
pillow
openpyxl
requests
pandas
